template types ainverseoperator
