+++ tests/stokes_solver_fail/screen-output
@@ -30,7 +30,9 @@ Additional information:
     void dealii::SolverFGMRES<VectorType>::solve(const MatrixType&,
     VectorType&, const VectorType&, const PreconditionerType&) [with
     MatrixType = aspect::internal::StokesBlock; PreconditionerType =
-    aspect::internal::BlockSchurPreconditioner<aspect::internal::InverseVelocityBlock<dealii::TrilinosWrappers::PreconditionAMG>,
+    aspect::internal::BlockSchurPreconditioner<aspect::internal::InverseVelocityBlock<dealii::TrilinosWrappers::PreconditionAMG,
+    dealii::TrilinosWrappers::MPI::Vector,
+    dealii::TrilinosWrappers::SparseMatrix>,
     aspect::internal::SchurComplementOperator,
     dealii::TrilinosWrappers::SparseMatrix,
     dealii::TrilinosWrappers::MPI::BlockVector>; VectorType =
